% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.causal_competing_risks_contrast}
\alias{plot.causal_competing_risks_contrast}
\title{Plot Effect-Over-Time Curves  (PLACEHOLDER — to be rewritten)}
\usage{
\method{plot}{causal_competing_risks_contrast}(x, ...)
}
\description{
Plot Effect-Over-Time Curves  (PLACEHOLDER — to be rewritten)
}
\keyword{internal}
