% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.causal_competing_risks_diagnostic}
\alias{plot.causal_competing_risks_diagnostic}
\title{Plot Weight Diagnostics  (PLACEHOLDER — deferred)}
\usage{
\method{plot}{causal_competing_risks_diagnostic}(x, ...)
}
\description{
Plot Weight Diagnostics  (PLACEHOLDER — deferred)
}
\keyword{internal}
