% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.causal_competing_risks_fit}
\alias{print.causal_competing_risks_fit}
\title{Print a causal_competing_risks Object}
\usage{
\method{print}{causal_competing_risks_fit}(x, ...)
}
\arguments{
\item{x}{A \code{"causal_competing_risks_fit"} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a causal_competing_risks Object
}
